+++ src/test/java/com/microsoft/store/partnercenter/models/subscriptions/ConversionErrorCodeTest.java
@@ -16,12 +16,8 @@ class ConversionErrorCodeTest
 {
     private static final String OTHER_VALUE = "other";
     private static final String CONVERSIONS_NOT_FOUND_VALUE = "conversions_not_found";
-    private static final String PENDING_VALUE = "pending";
-    private static final String FAILED_VALUE = "failed";
     private static final String OTHER_JSON = '"' + OTHER_VALUE + '"';
     private static final String CONVERSIONS_NOT_FOUND_JSON = '"' + CONVERSIONS_NOT_FOUND_VALUE + '"';
-    private static final String PENDING_JSON = '"' + PENDING_VALUE + '"';
-    private static final String FAILED_JSON = '"' + FAILED_VALUE + '"';
 
     private ObjectMapper jsonConverter;
 
@@ -36,8 +32,6 @@ class ConversionErrorCodeTest
     {
         assertEquals(ConversionErrorCode.OTHER, jsonConverter.readValue(OTHER_JSON, ConversionErrorCode.class));
         assertEquals(ConversionErrorCode.CONVERSIONS_NOT_FOUND, jsonConverter.readValue(CONVERSIONS_NOT_FOUND_JSON, ConversionErrorCode.class));
-        assertEquals(ConversionErrorCode.PENDING, jsonConverter.readValue(PENDING_JSON, ConversionErrorCode.class));
-        assertEquals(ConversionErrorCode.FAILED, jsonConverter.readValue(FAILED_JSON, ConversionErrorCode.class));
     }
 
     @Test
@@ -45,8 +39,6 @@ class ConversionErrorCodeTest
     {
         assertEquals(OTHER_JSON, jsonConverter.writeValueAsString(ConversionErrorCode.OTHER));
         assertEquals(CONVERSIONS_NOT_FOUND_JSON, jsonConverter.writeValueAsString(ConversionErrorCode.CONVERSIONS_NOT_FOUND));
-        assertEquals(PENDING_JSON, jsonConverter.writeValueAsString(ConversionErrorCode.PENDING));
-        assertEquals(FAILED_JSON, jsonConverter.writeValueAsString(ConversionErrorCode.FAILED));
     }
 
     @Test
@@ -54,7 +46,5 @@ class ConversionErrorCodeTest
     {
         assertEquals(OTHER_VALUE, ConversionErrorCode.OTHER.toString());
         assertEquals(CONVERSIONS_NOT_FOUND_VALUE, ConversionErrorCode.CONVERSIONS_NOT_FOUND.toString());
-        assertEquals(PENDING_VALUE, ConversionErrorCode.PENDING.toString());
-        assertEquals(FAILED_VALUE, ConversionErrorCode.FAILED.toString());
     }
 }